:tractor: flatten concentrated files
