+++ moban/plugins/library.py
@@ -1,5 +1,6 @@
-from moban import constants
 from lml.plugin import PluginManager
+
+from moban import constants
 
 
 class LibraryManager(PluginManager):
