+++ tests/__init__.py
@@ -1,0 +1,5 @@
+from moban.main import load_engine_factory_and_engines
+
+
+def setup():
+    load_engine_factory_and_engines()
