+++ tests/test_jinja2_extensions.py
@@ -2,9 +2,9 @@ import os
 
 from nose.tools import eq_
 
-from moban.plugins import BaseEngine
 from moban.jinja2.engine import Engine
 from moban.jinja2.extensions import jinja_global
+from moban.plugins.template_engine import TemplateEngine
 
 
 def test_globals():
@@ -12,7 +12,7 @@ def test_globals():
     test_dict = dict(hello="world")
     jinja_global("test", test_dict)
     path = os.path.join("tests", "fixtures", "globals")
-    engine = BaseEngine([path], path, Engine)
+    engine = TemplateEngine([path], path, Engine)
     engine.render_to_file("basic.template", "basic.yml", output)
     with open(output, "r") as output_file:
         content = output_file.read()
