+++ tests/test_template.py
@@ -5,8 +5,10 @@ from mock import patch
 from moban.plugins import ENGINES
 from moban.definitions import TemplateTarget
 
+MODULE = "moban.plugins.template_engine"
 
-@patch("moban.plugins.BaseEngine._render_with_finding_data_first")
+
+@patch(MODULE + ".TemplateEngine._render_with_finding_data_first")
 def test_do_templates_1(_do_templates_with_more_shared_data):
     jobs = [
         TemplateTarget("1.template", "data.yml", "1.output"),
@@ -31,7 +33,7 @@ def test_do_templates_1(_do_templates_with_more_shared_data):
         os.unlink(".moban.hashes")
 
 
-@patch("moban.plugins.BaseEngine._render_with_finding_template_first")
+@patch(MODULE + ".TemplateEngine._render_with_finding_template_first")
 def test_do_templates_2(_do_templates_with_more_shared_templates):
     jobs = [
         TemplateTarget("1.template", "data1.yml", "1.output"),
