+++ tests/test_yaml_loader.py
@@ -2,8 +2,8 @@ import os
 
 from nose.tools import eq_, raises
 
-from moban.data_loaders.manager import load_data
 from moban.data_loaders.yaml import open_yaml
+from moban.data_loaders.manager import load_data
 
 
 def test_simple_yaml():
